requirements.txt
alembic==1.16.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
billiard==4.2.1
celery==5.5.3
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
exceptiongroup==1.3.0
fastapi==0.115.12
google-play-scraper @ git+https://github.com/GilangIRSYD/google-play-scraper.git@266c1d9e098b679b87971cac7157397d8ce85f89
h11==0.16.0
idna==3.10
kombu==5.5.4
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.9
pydantic==2.11.5
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
redis==6.2.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.34.2
vine==5.1.0
wcwidth==0.2.13
